Move and rename GetCSharpSyntaxTree extension method
This moves GetCSharpSyntaxTree from RazorCodeDocumentExtensions to be a static method on DocumentSnapshot. This should help keep it from being called in co-hosting scenarios.
